Created the ball object.
