+++ src/Game.java
@@ -5,48 +5,29 @@ import java.awt.geom.Ellipse2D;
 
 public class Game extends JPanel{
 	
-	int x = 0;
-	int y = 0;
-	int xa = 1;
-	int ya = 1;
+	static Ball ball = new Ball();
 	
-	public void moveBall() {
+	private void move() {
 		
-		if (x + xa < 0) {
-			
-			xa = 1;
-			
-		}
-		if (x + xa > getWidth() - 30) {
-			
-			xa = -1;
-			
-		}
-		
-		if (y + ya < 0) {
-			
-			ya = 1;
-			
-		}
-		if (y + ya > getHeight() - 30) {
-			
-			ya = -1;
-			
-		}
-		
-		x = x + xa;
-		y = y + ya;
+		ball.moveBall();
 		
 	}
 	
-	@Override
-	public void paint (Graphics g) {
+	public static void StartGame() throws InterruptedException {
 		
-		super.paint(g);
-		Graphics2D g2d = (Graphics2D) g;
-		g2d.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
+		JFrame frame = new JFrame("Arkanoid");
+		frame.add(ball);
+		frame.setSize(600, 600);
+		frame.setVisible(true);
+		frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		
-		g2d.fillOval(x, y, 30, 30);
+		while (true) {
+			
+			ball.moveBall();
+			ball.repaint();
+			Thread.sleep(5);
+			
+		}
 		
 	}
