+++ src/main.java
@@ -12,20 +12,7 @@ public class main {
 	
 	public void principal() throws InterruptedException {
 		
-		JFrame frame = new JFrame("Arkanoid");
-		Game game = new Game();
-		frame.add(game);
-		frame.setSize(600, 600);
-		frame.setVisible(true);
-		frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-		
-		while (true) {
-			
-			game.moveBall();
-			game.repaint();
-			Thread.sleep(5);
-			
-		}
+		Game.StartGame();
 		
 	}
